Backend/requirements.txt
fastapi
uvicorn
pymongo[srv]
pydantic
python-multipart
python-dotenv
passlib[bcrypt]
bcrypt
numpy
scikit-learn
